set LFMainactivity as main page
